serasa/requirements.txt
-i https://pypi.org/simple
babel==2.17.0
blinker==1.9.0 ; python_version >= '3.9'
cachetools==5.5.2 ; python_version >= '3.7'
certifi==2025.1.31 ; python_version >= '3.6'
charset-normalizer==3.4.1 ; python_version >= '3.7'
click==8.1.8 ; python_version >= '3.7'
cloudevents==1.11.0
deprecation==2.1.0
flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.2 ; python_version >= '3.7'
google-api-python-client==2.164.0
google-auth==2.38.0 ; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2 ; python_version >= '3.7'
gunicorn==23.0.0 ; platform_system != 'Windows'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10 ; python_version >= '3.6'
itsdangerous==2.2.0 ; python_version >= '3.8'
jinja2==3.1.6 ; python_version >= '3.7'
markupsafe==3.0.2 ; python_version >= '3.9'
oauthlib==3.2.2 ; python_version >= '3.6'
packaging==24.2 ; python_version >= '3.8'
proto-plus==1.26.1 ; python_version >= '3.7'
protobuf==6.30.1 ; python_version >= '3.9'
pyasn1==0.6.1 ; python_version >= '3.8'
pyasn1-modules==0.4.1 ; python_version >= '3.8'
pyparsing==3.2.1 ; python_version >= '3.1'
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0 ; python_version >= '3.4'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.3.0 ; python_version >= '3.9'
watchdog==6.0.0 ; python_version >= '3.9'
werkzeug==3.1.3 ; python_version >= '3.9'
